Couple Things
Home / Guide has been split from the 'newStyle' file, mostly just because I wanted to try something to fix that positioning issue and it is a very difficult file to look through.

Add Post has had a complete makeover. It is now using Quill.JS for its text editor. It works as expected.

Navigation.
The highlight of these changes, it works as expected.